+++ packages/server/src/notifier/createNotifications.ts
@@ -1,32 +1,29 @@
-import { Prisma } from '@prisma/client'
+import { Prisma, Subscription } from '@prisma/client'
 import { request } from 'graphql-request'
 import { clone, groupBy, isEqual, isObject, mapValues } from 'lodash'
 import { info, verbose, warn } from 'npmlog'
 
 import { QUERY_NODE_ENDPOINT, STARTING_BLOCK } from '@/common/config'
 import { prisma } from '@/common/prisma'
-import { GetNotificationEventsDocument } from '@/common/queries'
+import { GetCurrentRolesDocument, GetNotificationEventsDocument } from '@/common/queries'
 import { count, getTypename } from '@/common/utils'
 
 import { toNotificationEvents } from './model/event'
 import { notificationsFromEvent } from './model/notifications'
 import { filterSubscriptions } from './model/subscriptionFilters'
-
-interface ProgressDoc {
-  block: number
-  eventIds: string[]
-}
-const isProgressDoc = (consumed: any): consumed is ProgressDoc => isObject(consumed)
-const defaultProgress: ProgressDoc = { block: STARTING_BLOCK, eventIds: [] }
-const PROGRESS_KEY = { key: 'Progress' }
+import { Notification, NotificationEvent, PotentialNotif, ProgressDocument } from './types'
 
 export const createNotifications = async (): Promise<void> => {
   // Check the last block that where processed
   const { value: initialProgress } = (await prisma.store.findUnique({ where: PROGRESS_KEY })) ?? {}
-  const progress: ProgressDoc =
-    isProgressDoc(initialProgress) && initialProgress.block > STARTING_BLOCK ? clone(initialProgress) : defaultProgress
+  const progress: ProgressDocument =
+    isProgressDocument(initialProgress) && initialProgress.block > STARTING_BLOCK
+      ? clone(initialProgress)
+      : defaultProgress
 
   const allMembers = (await prisma.member.findMany()).map(({ id, receiveEmails }) => ({ id, receiveEmails }))
+
+  const qnRoles = await request(QUERY_NODE_ENDPOINT, GetCurrentRolesDocument)
 
   /* eslint-disable-next-line no-constant-condition */
   while (true) {
@@ -39,33 +36,40 @@ export const createNotifications = async (): Promise<void> => {
 
     // Fetch events from the query nodes and break if non are found
     const qnVariables = { from: progress.block, exclude: progress.eventIds }
-    const qnData = await request(QUERY_NODE_ENDPOINT, GetNotificationEventsDocument, qnVariables)
+    const qnEvents = await request(QUERY_NODE_ENDPOINT, GetNotificationEventsDocument, qnVariables)
     info(
       'QN events',
-      `Received ${qnData.events.length} new events`,
-      mapValues(groupBy(qnData.events, getTypename), count),
+      `Received ${qnEvents.events.length} new events`,
+      mapValues(groupBy(qnEvents.events, getTypename), count),
       `from block ${progress.block} onward excluding`,
       progress.eventIds
     )
 
-    if (qnData.events.length === 0) break
+    if (qnEvents.events.length === 0) break
 
     // Generate the potential notification based on the query nodes data
-    const events = await Promise.all(qnData.events.map(toNotificationEvents(allMembers.map(({ id }) => id))))
+    const events: NotificationEvent[] = await Promise.all(
+      qnEvents.events.map(
+        toNotificationEvents(
+          allMembers.map(({ id }) => id),
+          qnRoles
+        )
+      )
+    )
 
     // Update the progress
     progress.block = Math.max(progress.block, ...events.map((event) => event.inBlock))
     progress.eventIds = events.flatMap((event) => (event.inBlock === progress.block ? event.id : []))
 
-    const potentialNotifs = events.flatMap((event) => event.potentialNotifications)
+    const potentialNotifs: PotentialNotif[] = events.flatMap((event) => event.potentialNotifications)
     if (potentialNotifs.length === 0) continue
 
     // Fetch subscription related to the events
     const subscriptionFilter = { OR: filterSubscriptions(potentialNotifs) }
-    const subscriptions = await prisma.subscription.findMany({ where: subscriptionFilter })
+    const subscriptions: Subscription[] = await prisma.subscription.findMany({ where: subscriptionFilter })
 
     // Create and save new notifications
-    const notifications = events.flatMap(notificationsFromEvent(subscriptions, allMembers))
+    const notifications: Notification[] = events.flatMap(notificationsFromEvent(subscriptions, allMembers))
     info('New notifications', 'Saving', notifications.length, 'new notifications')
     verbose(
       'New notifications',
@@ -79,3 +83,7 @@ export const createNotifications = async (): Promise<void> => {
     }
   }
 }
+
+const isProgressDocument = (consumed: any): consumed is ProgressDocument => isObject(consumed)
+const defaultProgress: ProgressDocument = { block: STARTING_BLOCK, eventIds: [] }
+const PROGRESS_KEY = { key: 'Progress' }