+++ packages/server/src/notifier/model/event/utils/types.ts
@@ -1,29 +1,12 @@
 import { GetNotificationEventsQuery } from '@/common/queries'
 import { DocWithFragments } from '@/common/utils/types'
 import { EntitySubscriptionKind, GeneralSubscriptionKind } from '@/notifier/model/subscriptionKinds'
+import { EntityPotentialNotif, GeneralPotentialNotif, NotificationEvent } from '@/notifier/types'
 
 export type ImplementedQNEvent = DocWithFragments<Required<GetNotificationEventsQuery['events'][0]>>
 type QNEvent<T extends ImplementedQNEvent['__typename']> = { __typename: T } & ImplementedQNEvent
 
-type GeneralEventParams = {
-  kind: GeneralSubscriptionKind
-  relatedMembers: 'ANY' | { ids: number[] }
-  isDefault: boolean
-}
-type EntityEventParams = { kind: EntitySubscriptionKind; relatedEntityId: string }
-export type GeneralPotentialNotif = { priority: number } & GeneralEventParams
-export type EntitiyPotentialNotif = { priority: number } & EntityEventParams
-
-export type PotentialNotif = GeneralPotentialNotif | EntitiyPotentialNotif
-
-type PartialNotif = Omit<GeneralPotentialNotif, 'priority'> | Omit<EntitiyPotentialNotif, 'priority'>
-
-export interface NotificationEvent {
-  id: string
-  inBlock: number
-  entityId: string
-  potentialNotifications: PotentialNotif[]
-}
+type PartialNotif = Omit<GeneralPotentialNotif, 'priority'> | Omit<EntityPotentialNotif, 'priority'>
 
 export interface NotifsBuilder {
   generalEvent: (kind: GeneralSubscriptionKind, members: 'ANY' | (number | string)[]) => PartialNotif | []
@@ -37,7 +20,8 @@ export type BuildEvents = (
   build: (b: NotifsBuilder) => (PartialNotif | [])[]
 ) => NotificationEvent
 
-export type NotifEventFromQNEvent<T extends ImplementedQNEvent['__typename']> = (
+export type NotifEventFromQNEvent<T extends ImplementedQNEvent['__typename'], Args extends any[] = []> = (
   event: QNEvent<T>,
-  buildEvents: BuildEvents
+  buildEvents: BuildEvents,
+  ...args: Args
 ) => Promise<NotificationEvent>